LPS-53386 Extract TestRule logic into TestCallback
