+++ portal-impl/test/integration/com/liferay/portlet/documentlibrary/webdav/WebDAVEnvironmentConfigTestRule.java
@@ -14,56 +14,20 @@
 
 package com.liferay.portlet.documentlibrary.webdav;
 
-import com.liferay.portal.kernel.util.Tuple;
-import com.liferay.portal.kernel.webdav.methods.Method;
-import com.liferay.portal.test.rule.MainServletTestRule;
-
-import javax.servlet.http.HttpServletResponse;
-
-import org.junit.Assert;
-import org.junit.runner.Description;
+import com.liferay.portal.kernel.test.rule.BaseTestRule;
 
 /**
  * @author Miguel Pastor
  * @author Shuyang Zhou
  */
-public class WebDAVEnvironmentConfigTestRule extends MainServletTestRule {
+public class WebDAVEnvironmentConfigTestRule
+	extends BaseTestRule<Object, Object> {
 
 	public static final WebDAVEnvironmentConfigTestRule INSTANCE =
 		new WebDAVEnvironmentConfigTestRule();
 
-	@Override
-	protected void afterClass(Description description, Object object) {
-		_baseWebDAVTestCase.service(Method.DELETE, "", null, null);
-
-		super.afterClass(description, object);
-	}
-
-	@Override
-	protected Object beforeClass(Description description) {
-		Object object = super.beforeClass(description);
-
-		Tuple tuple = _baseWebDAVTestCase.service(Method.MKCOL, "", null, null);
-
-		int statusCode = BaseWebDAVTestCase.getStatusCode(tuple);
-
-		if (statusCode == HttpServletResponse.SC_METHOD_NOT_ALLOWED) {
-			_baseWebDAVTestCase.service(Method.DELETE, "", null, null);
-
-			tuple = _baseWebDAVTestCase.service(Method.MKCOL, "", null, null);
-
-			statusCode = BaseWebDAVTestCase.getStatusCode(tuple);
-
-			Assert.assertEquals(HttpServletResponse.SC_CREATED, statusCode);
-		}
-
-		return object;
-	}
-
 	private WebDAVEnvironmentConfigTestRule() {
+		super(WebDAVEnvironmentConfigTestCallback.INSTANCE);
 	}
-
-	private final BaseWebDAVTestCase _baseWebDAVTestCase =
-		new BaseWebDAVTestCase();
 
 }
