+++ portal-test-internal/src/com/liferay/portal/test/rule/LogAssertionAppender.java
@@ -15,6 +15,7 @@
 package com.liferay.portal.test.rule;
 
 import com.liferay.portal.kernel.util.StringBundler;
+import com.liferay.portal.test.rule.callback.LogAssertionTestCallback;
 
 import org.apache.log4j.AppenderSkeleton;
 import org.apache.log4j.Level;
@@ -55,7 +56,7 @@ public class LogAssertionAppender extends AppenderSkeleton {
 			ThrowableInformation throwableInformation =
 				loggingEvent.getThrowableInformation();
 
-			LogAssertionTestRule.caughtFailure(
+			LogAssertionTestCallback.caughtFailure(
 				new AssertionError(
 					sb.toString(), throwableInformation.getThrowable()));
 		}
