+++ portal-test-internal/src/com/liferay/portal/test/rule/LogAssertionHandler.java
@@ -15,6 +15,7 @@
 package com.liferay.portal.test.rule;
 
 import com.liferay.portal.kernel.util.StringBundler;
+import com.liferay.portal.test.rule.callback.LogAssertionTestCallback;
 
 import java.util.logging.Handler;
 import java.util.logging.Level;
@@ -50,7 +51,7 @@ public class LogAssertionHandler extends Handler {
 			sb.append(", message=");
 			sb.append(logRecord.getMessage());
 
-			LogAssertionTestRule.caughtFailure(
+			LogAssertionTestCallback.caughtFailure(
 				new AssertionError(sb.toString(), logRecord.getThrown()));
 		}
 	}
