+++ portal-test-internal/src/com/liferay/portal/test/rule/MainServletTestRule.java
@@ -14,24 +14,8 @@
 
 package com.liferay.portal.test.rule;
 
-import com.liferay.portal.kernel.log.Log;
-import com.liferay.portal.kernel.log.LogFactoryUtil;
-import com.liferay.portal.kernel.search.SearchEngineUtil;
-import com.liferay.portal.kernel.servlet.ServletContextPool;
 import com.liferay.portal.kernel.test.rule.BaseTestRule;
-import com.liferay.portal.kernel.test.util.TestPropsValues;
-import com.liferay.portal.kernel.util.StringPool;
-import com.liferay.portal.service.test.ServiceTestUtil;
-import com.liferay.portal.servlet.MainServlet;
-import com.liferay.portal.test.mock.AutoDeployMockServletContext;
-
-import javax.servlet.ServletException;
-
-import org.junit.runner.Description;
-
-import org.springframework.core.io.FileSystemResourceLoader;
-import org.springframework.mock.web.MockServletConfig;
-import org.springframework.mock.web.MockServletContext;
+import com.liferay.portal.test.rule.callback.MainServletTestCallback;
 
 /**
  * @author Miguel Pastor
@@ -43,53 +27,7 @@ public class MainServletTestRule extends BaseTestRule<Object, Object> {
 		new MainServletTestRule();
 
 	protected MainServletTestRule() {
+		super(MainServletTestCallback.INSTANCE);
 	}
-
-	@Override
-	protected void afterClass(Description description, Object object) {
-		ServiceTestUtil.destroyServices();
-
-		try {
-			SearchEngineUtil.removeCompany(TestPropsValues.getCompanyId());
-		}
-		catch (Exception e) {
-			_log.error(e, e);
-		}
-	}
-
-	@Override
-	protected Object beforeClass(Description description) {
-		ServiceTestUtil.initServices();
-
-		ServiceTestUtil.initPermissions();
-
-		if (_mainServlet == null) {
-			MockServletContext mockServletContext =
-				new AutoDeployMockServletContext(
-					new FileSystemResourceLoader());
-
-			ServletContextPool.put(StringPool.BLANK, mockServletContext);
-
-			MockServletConfig mockServletConfig = new MockServletConfig(
-				mockServletContext);
-
-			_mainServlet = new MainServlet();
-
-			try {
-				_mainServlet.init(mockServletConfig);
-			}
-			catch (ServletException se) {
-				throw new RuntimeException(
-					"The main servlet could not be initialized");
-			}
-		}
-
-		return null;
-	}
-
-	private static final Log _log = LogFactoryUtil.getLog(
-		MainServletTestRule.class);
-
-	private static MainServlet _mainServlet;
 
 }
