+++ portal-test/src/com/liferay/portal/kernel/test/rule/SynchronousMailTestRule.java
@@ -14,51 +14,22 @@
 
 package com.liferay.portal.kernel.test.rule;
 
-import com.liferay.portal.kernel.messaging.DestinationNames;
-import com.liferay.portal.kernel.test.util.MailServiceTestUtil;
-
-import org.junit.runner.Description;
+import com.liferay.portal.kernel.test.rule.callback.SynchronousDestinationTestCallback.SyncHandler;
+import com.liferay.portal.kernel.test.rule.callback.SynchronousMailTestCallback;
 
 /**
  * @author Manuel de la Peña
  * @author Roberto Díaz
  * @author Shuyang Zhou
  */
-public class SynchronousMailTestRule extends SynchronousDestinationTestRule {
+public class SynchronousMailTestRule
+	extends BaseTestRule<SyncHandler, SyncHandler> {
 
 	public static final SynchronousMailTestRule INSTANCE =
 		new SynchronousMailTestRule();
 
-	@Override
-	protected void afterClass(
-		Description description, SyncHandler syncHandler) {
-
-		syncHandler.restorePreviousSync();
-
-		MailServiceTestUtil.stop();
-	}
-
-	@Override
-	protected void afterMethod(
-		Description description, SyncHandler syncHandler) {
-
-		super.afterMethod(description, syncHandler);
-
-		MailServiceTestUtil.clearMessages();
-	}
-
-	@Override
-	protected SyncHandler beforeClass(Description description) {
-		MailServiceTestUtil.start();
-
-		SyncHandler syncHandler = new SyncHandler();
-
-		syncHandler.replaceDestination(DestinationNames.MAIL);
-
-		return syncHandler;
-	}
-
 	private SynchronousMailTestRule() {
+		super(SynchronousMailTestCallback.INSTANCE);
 	}
 
 }
